[#noissue] Cleanup TimeCount
